refactor: Change directory name to plural
