+++ src/app/posts/[id]/page.tsx
